test: add `3D` cases with `.mat` data
Add test cases to check the behavior of 3D structure with all ones
matrix and `rand()` matrix, where the test data is created with
`for-loops` using the MATLAB implementation and all of the input,
fractional order and their corresponding transform in each 3-dimensions
are saved to `.mat` file. In the provided test cases, these imported and
tested with the implemented n-dim `fracF()`.