feat: add `dfrft` layer with fixed eigenvectors
For an input that does not change size (which is usually the
case for almost every architecture), it is inefficient to calculate
the fixed eigenvectors of DFT matrix for every forward case.

Therefore, we provide a FixEvecs version of the DFrFTLayer to
calculate the new DFrFT matrix from fixed eigenvectors. This
version approximately two times faster than the original.